Updated icons and screen content
