+++ client/package.json
@@ -35,6 +35,7 @@
     "react": "18.3.1",
     "react-dom": "^18.3.1",
     "react-native": "0.76.7",
+    "react-native-elements": "^3.4.3",
     "react-native-gesture-handler": "~2.20.2",
     "react-native-qrcode-svg": "^6.3.15",
     "react-native-radio-buttons-group": "^3.1.0",
