+++ client/src/screens/leaderboards.tsx
@@ -1,10 +1,18 @@
 import React from "react";
 import { View, Text, StyleSheet } from "react-native";
+import { useNavigation, NavigationProp } from "@react-navigation/native";
+import { RootStackParamList } from "@/types/types"; 
+import Navbar from '@/components/navbar';
+import Topbar from '@/components/topbar';
 
 export default function Leaderboard() {
+    const navigation = useNavigation<NavigationProp<RootStackParamList>>();
+
     return (
         <View style={styles.container}>
+            <Topbar />
             <Text>Leaderboard</Text>
+            <Navbar navigation={navigation} />
         </View>
     );
 }
